refactor: npc to be an abstract class and so a class per npc type
