+++ src/servers/ZoneServer2016/entities/npc.ts
@@ -16,33 +16,19 @@ import { ZoneServer2016 } from "../zoneserver";
 import { BaseFullCharacter } from "./basefullcharacter";
 import { ZoneClient2016 } from "../classes/zoneclient";
 import {
-  chance,
-  generateRandomGuid,
   getCurrentServerTimeWrapper,
   getDistance,
-  logClientActionToMongo,
-  randomIntFromInterval
+  logClientActionToMongo
 } from "../../../utils/utils";
 import { DB_COLLECTIONS, KILL_TYPE } from "../../../utils/enums";
-import {
-  Items,
-  MaterialTypes,
-  MeleeTypes,
-  ModelIds,
-  NpcIds,
-  PositionUpdateType,
-  StringIds
-} from "../models/enums";
+import { Items, MeleeTypes, NpcIds, PositionUpdateType } from "../models/enums";
 import { CommandInteractionString } from "types/zone2016packets";
 import { BaseEntity } from "./baseentity";
 import { ChallengeType } from "../managers/challengemanager";
 import { ProjectileEntity } from "./projectileentity";
-import { Lootbag } from "../entities/lootbag";
-import { LoadoutContainer } from "../classes/loadoutcontainer";
-import { ZombieInstance, createZombie } from "../jsms/zombie.jsm";
-import { getRandomZombieLoadout } from "../data/loadouts";
+import { ZombieInstance } from "../jsms/zombie.jsm";
 
-export class Npc extends BaseFullCharacter {
+export abstract class Npc extends BaseFullCharacter {
   health: number;
   npcRenderDistance = 100;
   spawnerId: number;
@@ -79,10 +65,11 @@ export class Npc extends BaseFullCharacter {
     return this.deathTime == 0;
   }
   server: ZoneServer2016;
-  npcMeleeDamage: number;
+  npcMeleeDamage: number = 0;
   zombieFsm?: ZombieInstance;
   currentAnimation = "";
   isSelected: boolean = false;
+
   constructor(
     characterId: string,
     transientId: number,
@@ -96,48 +83,22 @@ export class Npc extends BaseFullCharacter {
     this.positionUpdateType = PositionUpdateType.MOVABLE;
     this.spawnerId = spawnerId;
     this.health = 10000;
-    this.initNpcData();
     this.server = server;
     if (!process.env.DISABLE_AI && this.server.aiEnabled) {
       this.navAgent = this.server.navManager.createAgent(this.state.position);
     }
-    switch (actorModelId) {
-      case ModelIds.ZOMBIE_FEMALE_WALKER:
-      case ModelIds.ZOMBIE_MALE_WALKER:
-        this.materialType = MaterialTypes.ZOMBIE;
-        this.npcMeleeDamage = 2000;
-
-        if (!process.env.DISABLE_AI && this.server.aiEnabled) {
-          this.zombieFsm = createZombie(this, server);
-        }
-        for (const entry of getRandomZombieLoadout()) {
-          this.equipItem(server, server.generateItem(entry.item), false);
-        }
-        break;
-      case ModelIds.ZOMBIE_SCREAMER:
-        this.materialType = MaterialTypes.ZOMBIE;
-        this.npcMeleeDamage = 3000;
-        break;
-      case ModelIds.DEER:
-      case ModelIds.DEER_BUCK:
-        this.materialType = MaterialTypes.FLESH;
-        this.npcMeleeDamage = 0;
-        break;
-      case ModelIds.WOLF:
-        this.materialType = MaterialTypes.FLESH;
-        this.npcMeleeDamage = 2000;
-        break;
-      case ModelIds.BEAR:
-        this.materialType = MaterialTypes.FLESH;
-        this.npcMeleeDamage = 4000;
-        break;
-      default:
-        this.materialType = MaterialTypes.FLESH;
-        this.npcMeleeDamage = 0;
-        break;
-    }
     server.explosiveManager.addEntity(this);
   }
+
+  protected abstract addLoot(server: ZoneServer2016): void;
+  protected abstract onHarvest(
+    server: ZoneServer2016,
+    client: ZoneClient2016
+  ): void;
+  protected abstract buildInteractionString(
+    server: ZoneServer2016,
+    client: ZoneClient2016
+  ): void;
 
   setAnimation(animationName: string) {
     this.currentAnimation = animationName;
@@ -202,18 +163,7 @@ export class Npc extends BaseFullCharacter {
       this.zombieFsm?.destroyed();
       this.flags.knockedOut = 1;
 
-      // Custom lootbag for zombies
-      switch (this.actorModelId) {
-        case ModelIds.ZOMBIE_FEMALE_WALKER:
-        case ModelIds.ZOMBIE_MALE_WALKER:
-        case ModelIds.ZOMBIE_SCREAMER: {
-          this.addZombieLoot(server);
-          break;
-        }
-        default:
-          server.worldObjectManager.createLootbag(server, this);
-          break;
-      }
+      this.addLoot(server);
 
       if (client) {
         if (this.npcId === NpcIds.ZOMBIE) {
@@ -276,129 +226,6 @@ export class Npc extends BaseFullCharacter {
       );
       client.character.addCombatlogEntry(damageRecord);
     }
-  }
-
-  addZombieLoot(server: ZoneServer2016) {
-    const lootItems: any[] = [];
-
-    const wornLetters = [
-      Items.WORN_LETTER_CHURCH_PV,
-      Items.WORN_LETTER_LJ_PV,
-      Items.WORN_LETTER_MISTY_DAM,
-      Items.WORN_LETTER_RADIO,
-      Items.WORN_LETTER_RUBY_LAKE,
-      Items.WORN_LETTER_TOXIC_LAKE,
-      Items.WORN_LETTER_VILLAS,
-      Items.WORN_LETTER_WATER_TOWER
-    ];
-    // Worn letter (4% chance, up to 2)
-    if (chance(50)) {
-      const randomWornLetter =
-        wornLetters[randomIntFromInterval(0, wornLetters.length - 1)];
-      const wornLetterItem = server.generateItem(randomWornLetter, 1);
-      if (wornLetterItem) {
-        lootItems.push(wornLetterItem);
-      }
-    }
-
-    const GoodammoTypes = [
-      Items.AMMO_12GA,
-      Items.AMMO_223,
-      Items.AMMO_308,
-      Items.AMMO_762
-    ];
-    // Ammo (5% chance)
-    if (chance(50)) {
-      for (let i = 0; i < 2; i++) {
-        const randomAmmo =
-          GoodammoTypes[randomIntFromInterval(0, GoodammoTypes.length - 1)];
-        const ammoCount = randomIntFromInterval(1, 3);
-        const ammoItem = server.generateItem(randomAmmo, ammoCount);
-        if (ammoItem) {
-          lootItems.push(ammoItem);
-        }
-      }
-    }
-
-    const ammoTypes = [Items.AMMO_380, Items.AMMO_9MM, Items.AMMO_45];
-    // Ammo (10% chance)
-    if (chance(100)) {
-      for (let i = 0; i < ammoTypes.length - 1; i++) {
-        const randomAmmo =
-          ammoTypes[randomIntFromInterval(0, ammoTypes.length - 1)];
-        const ammoCount = randomIntFromInterval(1, 5);
-        const ammoItem = server.generateItem(randomAmmo, ammoCount);
-        if (ammoItem) {
-          lootItems.push(ammoItem);
-        }
-      }
-    }
-
-    const specialItems = [
-      Items.WEAPON_BOW_MAKESHIFT,
-      Items.BACKPACK_BLUE_ORANGE,
-      Items.CRUMPLED_NOTE,
-      Items.REFRIGERATOR_NOTE
-    ];
-    // Special item (15% chance)
-    if (chance(150)) {
-      for (let i = 0; i < specialItems.length - 1; i++) {
-        const randomSpecial =
-          specialItems[randomIntFromInterval(0, specialItems.length - 1)];
-        const specialItem = server.generateItem(randomSpecial, 1);
-        if (specialItem) {
-          lootItems.push(specialItem);
-        }
-      }
-    }
-
-    // Prototype item (1% chance)
-    const PrototypeItems = [
-      Items.PROTOTYPE_MECHANISM,
-      Items.PROTOTYPE_RECEIVER,
-      Items.PROTOTYPE_TRIGGER_ASSEMBLY
-    ];
-    if (chance(10)) {
-      const randomSpecial =
-        PrototypeItems[randomIntFromInterval(0, PrototypeItems.length - 1)];
-      const specialItem = server.generateItem(randomSpecial, 1);
-      if (specialItem) {
-        lootItems.push(specialItem);
-      }
-    }
-
-    // Cloth (80% chance)
-    if (chance(800)) {
-      const clothCount = randomIntFromInterval(1, 3);
-      const clothItem = server.generateItem(Items.CLOTH, clothCount);
-      if (clothItem) {
-        lootItems.push(clothItem);
-      }
-    }
-
-    // Only spawn lootbag if there is loot
-    if (lootItems.length === 0) return;
-
-    const characterId = generateRandomGuid();
-    const lootbag = new Lootbag(
-      characterId,
-      server.getTransientId(characterId),
-      ModelIds.LOOT_BAG_CLEAN,
-      new Float32Array([
-        this.state.position[0] + 0.7,
-        this.state.position[1],
-        this.state.position[2] + 0.7
-      ]),
-      new Float32Array([0, 0, 0, 0]),
-      server
-    );
-    const container = lootbag.getContainer();
-
-    for (const item of lootItems) {
-      server.addContainerItem(lootbag, item, container as LoadoutContainer);
-    }
-
-    server._lootbags[characterId] = lootbag;
   }
 
   OnFullCharacterDataRequest(server: ZoneServer2016, client: ZoneClient2016) {
@@ -467,73 +294,6 @@ export class Npc extends BaseFullCharacter {
     return server.deleteEntity(this.characterId, server._npcs);
   }
 
-  initNpcData() {
-    switch (this.actorModelId) {
-      case ModelIds.ZOMBIE_SCREAMER:
-        this.nameId = StringIds.BANSHEE;
-        this.npcId = NpcIds.ZOMBIE;
-        break;
-      case ModelIds.ZOMBIE_FEMALE_WALKER:
-      case ModelIds.ZOMBIE_MALE_WALKER:
-        this.nameId = StringIds.ZOMBIE_WALKER;
-        this.rewardItems = [
-          {
-            itemDefId: Items.BRAIN_INFECTED,
-            weight: 10
-          }
-        ];
-        this.npcId = NpcIds.ZOMBIE;
-        break;
-      case ModelIds.DEER_BUCK:
-      case ModelIds.DEER:
-        this.nameId = StringIds.DEER;
-        this.rewardItems = [
-          {
-            itemDefId: Items.MEAT_VENISON,
-            weight: 30
-          },
-          {
-            itemDefId: Items.ANIMAL_FAT,
-            weight: 20
-          },
-          {
-            itemDefId: Items.DEER_BLADDER,
-            weight: 10
-          }
-        ];
-        this.npcId = NpcIds.DEER;
-        break;
-      case ModelIds.WOLF:
-        this.nameId = StringIds.WOLF;
-        this.rewardItems = [
-          {
-            itemDefId: Items.MEAT_WOLF,
-            weight: 30
-          },
-          {
-            itemDefId: Items.ANIMAL_FAT,
-            weight: 20
-          }
-        ];
-        this.npcId = NpcIds.WOLF;
-        break;
-      case ModelIds.BEAR:
-        this.nameId = StringIds.BEAR;
-        this.rewardItems = [
-          {
-            itemDefId: Items.MEAT_BEAR,
-            weight: 40
-          },
-          {
-            itemDefId: Items.ANIMAL_FAT,
-            weight: 20
-          }
-        ];
-        this.npcId = NpcIds.BEAR;
-        break;
-    }
-  }
-
   OnPlayerSelect(server: ZoneServer2016, client: ZoneClient2016) {
     // Only one at a time
     if (this.isSelected) {
@@ -548,34 +308,7 @@ export class Npc extends BaseFullCharacter {
     const skinningKnife = client.character.getItemById(Items.SKINNING_KNIFE);
     if (!this.isAlive && skinningKnife) {
       server.utilizeHudTimer(client, this.nameId, 5000, 0, () => {
-        switch (this.actorModelId) {
-          case ModelIds.ZOMBIE_FEMALE_WALKER:
-          case ModelIds.ZOMBIE_MALE_WALKER:
-          case ModelIds.ZOMBIE_SCREAMER:
-            const emptySyringe = client.character.getItemById(
-              Items.SYRINGE_EMPTY
-            );
-            if (emptySyringe) {
-              client.character.lootContainerItem(
-                server,
-                server.generateItem(Items.SYRINGE_INFECTED_BLOOD)
-              );
-              server.removeInventoryItem(client.character, emptySyringe);
-              return;
-            }
-            this.triggerAwards(server, client, this.rewardItems);
-            break;
-          case ModelIds.DEER_BUCK:
-          case ModelIds.DEER:
-            this.triggerAwards(server, client, this.rewardItems);
-            break;
-          case ModelIds.BEAR:
-            this.triggerAwards(server, client, this.rewardItems);
-            break;
-          case ModelIds.WOLF:
-            this.triggerAwards(server, client, this.rewardItems);
-            break;
-        }
+        this.onHarvest(server, client);
         server.damageItem(client.character, skinningKnife, 200);
         server.deleteEntity(this.characterId, server._npcs);
       });
@@ -635,23 +368,7 @@ export class Npc extends BaseFullCharacter {
 
   OnInteractionString(server: ZoneServer2016, client: ZoneClient2016) {
     if (!this.isAlive && client.character.hasItem(Items.SKINNING_KNIFE)) {
-      switch (this.actorModelId) {
-        case ModelIds.ZOMBIE_FEMALE_WALKER:
-        case ModelIds.ZOMBIE_MALE_WALKER:
-        case ModelIds.ZOMBIE_SCREAMER:
-          if (client.character.hasItem(Items.SYRINGE_EMPTY)) {
-            this.sendInteractionString(server, client, StringIds.EXTRACT_BLOOD);
-            return;
-          }
-          this.sendInteractionString(server, client, StringIds.HARVEST);
-          break;
-        case ModelIds.DEER_BUCK:
-        case ModelIds.DEER:
-        case ModelIds.WOLF:
-        case ModelIds.BEAR:
-          this.sendInteractionString(server, client, StringIds.HARVEST);
-          break;
-      }
+      this.buildInteractionString(server, client);
     }
   }
 
@@ -669,10 +386,12 @@ export class Npc extends BaseFullCharacter {
       }
     );
   }
+
   stopMovement() {
     if (!this.navAgent) return;
     this.navAgent.resetMoveTarget();
   }
+
   goTo(position: Float32Array) {
     const dx = position[0] - this.state.position[0];
     const dz = position[2] - this.state.position[2];
